auth/requirements.txt
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.7
Django==3.2
djangorestframework==3.12.4
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
django-simple-history==2.12.0
-e git+https://github.com/altostratous/django-simple-history-field-track.git@dacaaedd212e57f17c9152d7dcbcb2e68c676c59#egg=django_simple_history_field_track
python-decouple==3.1
